Monitor changes to command history to avoid crash
When command history has stabilized, run as before, operating directly on the array of commands for each Tortoise instance.

When the command history is in flux, make a local copy of the array of command for each Tortoise instance so that the current thread doesn't get tripped up by changes made by other threads.

Locking threads is the more correct and elegant solution but seems to have an adverse effect on animation speed.